Add ability to extract values from PQL qury results
When users run PQL queries, it returns an array of hashes where each
hash is an item that matched the query. If users want to extract
information from the results, such as ipaddresses or certnames, the user
must iterate over the PQL results. However, this is highly discouraged
in Jenkins pipeline scripts.

This solution enables users to extract out information into a flat array
so they do not need to resort to iterating over the results.